+++ frontend/package.json
@@ -32,6 +32,7 @@
   "imports": {
     "#utils/*": "./src/utils/*.js",
     "#default-data/*": "./src/utils/default-data/*.js",
+    "#test-data/*": "./src/utils/test-data/*.js",
     "#fetch/*": "./src/utils/fetch/*.js",
     "#nonprofitPage/*": "./src/pages/nonprofit/*.jsx",
     "#refugeePage/*": "./src/pages/refugee/*.jsx",
