+++ frontend/src/components/Register/Register.jsx
@@ -72,7 +72,6 @@ function Register() {
     let allLocations = location.split("/");
     let ending = allLocations[allLocations.length - 1];
     let newPath = location.replace(ending, "login");
-    console.log(newPath);
     navigate(newPath);
   }
   return (
