+++ frontend/src/components/Service/Service.css
@@ -1,3 +1,7 @@
-/* .Service{
-
-} */
+.Service {
+  padding: 0 3% 0% 3%;
+  margin: 3%;
+  text-align: left;
+  border: var(--border-mid) solid 10px;
+  border-radius: 30px;
+}
