+++ frontend/src/pages/nonprofit/ViewServices/ViewServices.jsx
@@ -1,16 +1,25 @@
-import React from "react";
+import React, { use } from "react";
 import ReactDOM from "react-dom";
+import { useState, useEffect } from "react";
 import "./ViewServices.css";
 import PropTypes from "prop-types";
-import SearchFilters from "#components/SearchFilters/SearchFilters";
+
+// Other components
 import ServiceList from "#components/ServiceList/ServiceList";
+// util functions
+import { fetchAllServices } from "#fetch/serviceFetchUtils";
 
 function ViewServices({}) {
+  const [searchResults, setSearchResults] = useState([]);
+  useEffect(() => {
+    fetchAllServices().then((results) => {
+      setSearchResults(results.data);
+    });
+  });
   return (
     <div className="ViewServices">
-      <h3>ViewServices</h3>
-      <SearchFilters />
-      <ServiceList />
+      <h1>All Services</h1>
+      <ServiceList data={searchResults} />
     </div>
   );
 }
