+++ frontend/src/utils/default-data/serviceDefaultData.js
@@ -176,7 +176,7 @@ export const serviceSearchDefault = [
   },
   {
     id: "country_of_origin",
-    name: "Country of Origin*",
+    name: "Country of Origin",
     value: "",
     default: "Afghanistan",
     required: false,
@@ -184,7 +184,7 @@ export const serviceSearchDefault = [
   },
   {
     id: "language",
-    name: "Language Spoken*",
+    name: "Language Spoken",
     value: "",
     default: "Arabic",
     required: false,
@@ -192,7 +192,7 @@ export const serviceSearchDefault = [
   },
   {
     id: "date_entered",
-    name: "Date entered US*",
+    name: "Date entered US",
     value: "",
     default: "2025-01-01",
     required: false,
@@ -205,83 +205,5 @@ export const serviceSearchDefault = [
     default: "2",
     required: false,
     icon: MdOutlinePersonAddAlt,
-  },
-];
-export const serviceSearchTestData = [
-  {
-    id: 1,
-    name: "American Red Cross",
-    address: "2025 E St NW, Washington, DC 20006",
-    zipcode: "20006",
-    description:
-      "Humanitarian organization providing emergency assistance and disaster relief.",
-    phone_number: "(202) 303-4498",
-    email: "info@redcross.org",
-    phone: "(202) 303-4498",
-    website: "https://www.redcross.org/",
-    hours: "Monday - Friday, 9am - 5pm",
-    logo: "https://www.redcross.org/images/logo.png",
-    language: "English, Spanish",
-    date_needed: "Ongoing",
-    services_offered: "Disaster Relief, Blood Donations, First Aid Training",
-    restrictions: "None",
-    next_steps: "Visit website to volunteer or donate",
-  },
-  {
-    id: 2,
-    name: "Habitat for Humanity",
-    address: "322 W Lamar St, Americus, GA 31709",
-    zipcode: "31709",
-    description:
-      "Non-profit organization building affordable housing for low-income families.",
-    phone_number: "(229) 924-6935",
-    email: "info@habitat.org",
-    phone: "(229) 924-6935",
-    website: "https://www.habitat.org/",
-    hours: "Monday - Friday, 8am - 4:30pm",
-    logo: "https://www.habitat.org/images/logo.png",
-    language: "English",
-    date_needed: "Ongoing",
-    services_offered: "Home Building, Home Repair, Disaster Response",
-    restrictions: "Must meet income guidelines",
-    next_steps: "Visit website to apply for housing or volunteer",
-  },
-  {
-    id: 3,
-    name: "The Salvation Army",
-    address: "615 Slaters Ln, Alexandria, VA 22314",
-    zipcode: "22314",
-    description:
-      "International organization providing social services, disaster relief, and spiritual support.",
-    phone_number: "(703) 684-5500",
-    email: "info@salvationarmy.org",
-    phone: "(703) 684-5500",
-    website: "https://www.salvationarmy.org/",
-    hours: "Monday - Friday, 9am - 4pm",
-    logo: "https://www.salvationarmy.org/images/logo.png",
-    language: "English, Spanish",
-    date_needed: "Ongoing",
-    services_offered: "Food and Shelter, Disaster Relief, Youth Programs",
-    restrictions: "None",
-    next_steps: "Visit website to donate or volunteer",
-  },
-  {
-    id: 4,
-    name: "St. Jude Children's Research Hospital",
-    address: "262 Danny Thomas Pl, Memphis, TN 38105",
-    zipcode: "38105",
-    description:
-      "Pediatric cancer research hospital providing free medical care to children.",
-    phone_number: "(901) 595-3300",
-    email: "info@stjude.org",
-    phone: "(901) 595-3300",
-    website: "https://www.stjude.org/",
-    hours: "Monday - Friday, 8am - 5pm",
-    logo: "https://www.stjude.org/images/logo.png",
-    language: "English",
-    date_needed: "Ongoing",
-    services_offered: "Cancer Treatment, Research, Patient Care",
-    restrictions: "Children",
-    next_steps: "Visit website to donate or volunteer",
   },
 ];
