+++ frontend/src/utils/test-data/serviceTestData.js
@@ -1,0 +1,78 @@
+export const serviceSearchTestData = [
+  {
+    id: 1,
+    name: "American Red Cross",
+    address: "2025 E St NW, Washington, DC 20006",
+    distance: "10 miles",
+    description:
+      "Humanitarian organization providing emergency assistance and disaster relief.",
+    phone_number: "(202) 303-4498",
+    email: "info@redcross.org",
+    phone: "(202) 303-4498",
+    website: "https://www.redcross.org/",
+    hours: "Monday - Friday, 9am - 5pm",
+    logo: "https://www.redcross.org/images/logo.png",
+    language: "English, Spanish",
+    date_needed: "Ongoing",
+    services_offered: "Disaster Relief, Blood Donations, First Aid Training",
+    restrictions: "None",
+    next_steps: "Visit website to volunteer or donate",
+  },
+  {
+    id: 2,
+    name: "Habitat for Humanity",
+    address: "322 W Lamar St, Americus, GA 31709",
+    distance: "10 miles",
+    description:
+      "Non-profit organization building affordable housing for low-income families.",
+    phone_number: "(229) 924-6935",
+    email: "info@habitat.org",
+    phone: "(229) 924-6935",
+    website: "https://www.habitat.org/",
+    hours: "Monday - Friday, 8am - 4:30pm",
+    logo: "https://www.habitat.org/images/logo.png",
+    language: "English",
+    date_needed: "Ongoing",
+    services_offered: "Home Building, Home Repair, Disaster Response",
+    restrictions: "Must meet income guidelines",
+    next_steps: "Visit website to apply for housing or volunteer",
+  },
+  {
+    id: 3,
+    name: "The Salvation Army",
+    address: "615 Slaters Ln, Alexandria, VA 22314",
+    distance: "10 miles",
+    description:
+      "International organization providing social services, disaster relief, and spiritual support.",
+    phone_number: "(703) 684-5500",
+    email: "info@salvationarmy.org",
+    phone: "(703) 684-5500",
+    website: "https://www.salvationarmy.org/",
+    hours: "Monday - Friday, 9am - 4pm",
+    logo: "https://www.salvationarmy.org/images/logo.png",
+    language: "English, Spanish",
+    date_needed: "Ongoing",
+    services_offered: "Food and Shelter, Disaster Relief, Youth Programs",
+    restrictions: "None",
+    next_steps: "Visit website to donate or volunteer",
+  },
+  {
+    id: 4,
+    name: "St. Jude Children's Research Hospital",
+    address: "262 Danny Thomas Pl, Memphis, TN 38105",
+    distance: "10 miles",
+    description:
+      "Pediatric cancer research hospital providing free medical care to children.",
+    phone_number: "(901) 595-3300",
+    email: "info@stjude.org",
+    phone: "(901) 595-3300",
+    website: "https://www.stjude.org/",
+    hours: "Monday - Friday, 8am - 5pm",
+    logo: "https://www.stjude.org/images/logo.png",
+    language: "English",
+    date_needed: "Ongoing",
+    services_offered: "Cancer Treatment, Research, Patient Care",
+    restrictions: "Children",
+    next_steps: "Visit website to donate or volunteer",
+  },
+];
